+++ R/missing_package_linter.R
@@ -3,7 +3,7 @@
 #' @export
 missing_package_linter <- function(source_file) {
 
-  if (!length(source_file$parsed_content) || is.null(source_file$xml_parsed_content)) return(list())
+  if (!length(source_file$full_parsed_content) || is.null(source_file$xml_parsed_content)) return(list())
 
   xml <- source_file$xml_parsed_content
 
@@ -36,7 +36,7 @@ missing_package_linter <- function(source_file) {
       column_number = col1[[i]],
       type = "warning",
       message = sprintf("Package '%s' is not installed.", pkg_names[[i]]),
-      line = source_file$lines[line1[[i]]],
+      line = source_file$file_lines[line1[[i]]],
       ranges = list(c(col1[[i]], col2[[i]])),
       linter = "missing_package_linter"
     )