+++ NEWS.md
@@ -10,6 +10,11 @@
 * New `missing_package_linter()` (#536, #547, @renkun-ken)
 * New `namespace_linter()` (#548, #551, @renkun-ken)
 * Fix possible error on invalid XML produced by xmlparsedata (#559, #560, @renkun-ken)
+* Fix handling zero-length variable name error (#566, #567, @renkun-ken)
+* New `missing_argument_linter()` (#563, #565, @renkun-ken)
+* New `sprintf_linter()` (#544, #578, @renkun-ken)
+* Exclusions specified in the `.lintr` file are now relative to the location of that file 
+  and support excluding entire directories (#158, #438, @AshesITR)
 
 # lintr 2.0.1